Fix error type provided to showUrl callback on iOS
The existing code was providing an `NSError` to the callback for
`showUrl`. This ends up being `null` in the calling Javascript code
which results in an incomplete Promise that should have been
rejected.

This change simply creates a dictionary with the error information
so the Javascript code recognizes the error.